chore: 732 - rename index to trainers-api
